LPS-191342 Adapted to refactor of CongfigurationAdmin
